HSEARCH-3053 Remove the temporary package
