+++ backend/elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/work/builder/factory/impl/ElasticsearchWorkBuilderFactoryImpl.java
@@ -30,24 +30,24 @@ import org.hibernate.search.backend.elasticsearch.work.builder.impl.RefreshWorkB
 import org.hibernate.search.backend.elasticsearch.work.builder.impl.ScrollWorkBuilder;
 import org.hibernate.search.backend.elasticsearch.work.builder.impl.SearchWorkBuilder;
 import org.hibernate.search.backend.elasticsearch.work.impl.ElasticsearchSearchResultExtractor;
-import org.hibernate.search.backend.elasticsearch.work.real.impl.ClearScrollWork;
-import org.hibernate.search.backend.elasticsearch.work.real.impl.CloseIndexWork;
-import org.hibernate.search.backend.elasticsearch.work.real.impl.CountWork;
-import org.hibernate.search.backend.elasticsearch.work.real.impl.CreateIndexWork;
-import org.hibernate.search.backend.elasticsearch.work.real.impl.DeleteByQueryWork;
-import org.hibernate.search.backend.elasticsearch.work.real.impl.DeleteWork;
-import org.hibernate.search.backend.elasticsearch.work.real.impl.DropIndexWork;
-import org.hibernate.search.backend.elasticsearch.work.real.impl.ExplainWork;
-import org.hibernate.search.backend.elasticsearch.work.real.impl.FlushWork;
-import org.hibernate.search.backend.elasticsearch.work.real.impl.GetIndexSettingsWork;
-import org.hibernate.search.backend.elasticsearch.work.real.impl.IndexExistsWork;
-import org.hibernate.search.backend.elasticsearch.work.real.impl.IndexWork;
-import org.hibernate.search.backend.elasticsearch.work.real.impl.OpenIndexWork;
-import org.hibernate.search.backend.elasticsearch.work.real.impl.OptimizeWork;
-import org.hibernate.search.backend.elasticsearch.work.real.impl.PutIndexSettingsWork;
-import org.hibernate.search.backend.elasticsearch.work.real.impl.RefreshWork;
-import org.hibernate.search.backend.elasticsearch.work.real.impl.ScrollWork;
-import org.hibernate.search.backend.elasticsearch.work.real.impl.SearchWork;
+import org.hibernate.search.backend.elasticsearch.work.impl.ClearScrollWork;
+import org.hibernate.search.backend.elasticsearch.work.impl.CloseIndexWork;
+import org.hibernate.search.backend.elasticsearch.work.impl.CountWork;
+import org.hibernate.search.backend.elasticsearch.work.impl.CreateIndexWork;
+import org.hibernate.search.backend.elasticsearch.work.impl.DeleteByQueryWork;
+import org.hibernate.search.backend.elasticsearch.work.impl.DeleteWork;
+import org.hibernate.search.backend.elasticsearch.work.impl.DropIndexWork;
+import org.hibernate.search.backend.elasticsearch.work.impl.ExplainWork;
+import org.hibernate.search.backend.elasticsearch.work.impl.FlushWork;
+import org.hibernate.search.backend.elasticsearch.work.impl.GetIndexSettingsWork;
+import org.hibernate.search.backend.elasticsearch.work.impl.IndexExistsWork;
+import org.hibernate.search.backend.elasticsearch.work.impl.IndexWork;
+import org.hibernate.search.backend.elasticsearch.work.impl.OpenIndexWork;
+import org.hibernate.search.backend.elasticsearch.work.impl.OptimizeWork;
+import org.hibernate.search.backend.elasticsearch.work.impl.PutIndexSettingsWork;
+import org.hibernate.search.backend.elasticsearch.work.impl.RefreshWork;
+import org.hibernate.search.backend.elasticsearch.work.impl.ScrollWork;
+import org.hibernate.search.backend.elasticsearch.work.impl.SearchWork;
 
 import com.google.gson.JsonObject;
