+++ backend/elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/work/impl/DeleteWork.java
@@ -4,23 +4,24 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.backend.elasticsearch.work.real.impl;
+package org.hibernate.search.backend.elasticsearch.work.impl;
 
 import org.hibernate.search.backend.elasticsearch.client.spi.ElasticsearchRequest;
 import org.hibernate.search.backend.elasticsearch.client.spi.ElasticsearchResponse;
 import org.hibernate.search.backend.elasticsearch.util.spi.URLEncodedString;
-import org.hibernate.search.backend.elasticsearch.work.builder.impl.IndexWorkBuilder;
-import org.hibernate.search.backend.elasticsearch.work.impl.ElasticsearchWorkExecutionContext;
-import org.hibernate.search.backend.elasticsearch.work.real.accessor.impl.DefaultElasticsearchRequestSuccessAssessor;
+import org.hibernate.search.backend.elasticsearch.work.builder.impl.DeleteWorkBuilder;
 
 import com.google.gson.JsonObject;
 
 /**
  * @author Yoann Rodiere
  */
-public class IndexWork extends AbstractSimpleBulkableElasticsearchWork<Void> {
+public class DeleteWork extends AbstractSimpleBulkableElasticsearchWork<Void> {
 
-	public IndexWork(Builder builder) {
+	private static final ElasticsearchRequestSuccessAssessor SUCCESS_ASSESSOR =
+			DefaultElasticsearchRequestSuccessAssessor.builder().ignoreErrorStatuses( 404 ).build();
+
+	public DeleteWork(Builder builder) {
 		super( builder );
 	}
 
@@ -36,34 +37,30 @@ public class IndexWork extends AbstractSimpleBulkableElasticsearchWork<Void> {
 
 	public static class Builder
 			extends AbstractSimpleBulkableElasticsearchWork.Builder<Builder>
-			implements IndexWorkBuilder {
+			implements DeleteWorkBuilder {
 		private final URLEncodedString indexName;
 		private final URLEncodedString typeName;
 		private final URLEncodedString id;
 		private final String routingKey;
-		private final JsonObject document;
 
-		public Builder(URLEncodedString indexName, URLEncodedString typeName, URLEncodedString id, String routingKey, JsonObject document) {
-			super( indexName, DefaultElasticsearchRequestSuccessAssessor.INSTANCE );
+		public Builder(URLEncodedString indexName, URLEncodedString typeName, URLEncodedString id, String routingKey) {
+			super( indexName, SUCCESS_ASSESSOR );
 			this.indexName = indexName;
 			this.typeName = typeName;
 			this.id = id;
 			this.routingKey = routingKey;
-			this.document = document;
 		}
 
 		@Override
 		protected ElasticsearchRequest buildRequest() {
 			ElasticsearchRequest.Builder builder =
-					ElasticsearchRequest.put()
+					ElasticsearchRequest.delete()
 					.pathComponent( indexName )
 					.pathComponent( typeName )
-					.pathComponent( id )
+					.pathComponent( id );
 
-					// TODO avoid this param using a smart orchestrator
-					.param( "refresh", true )
-
-					.body( document );
+			// TODO avoid this param using a smart orchestrator
+			builder.param( "refresh", true );
 
 			if ( routingKey != null ) {
 				builder.param( "routing", routingKey );
@@ -74,20 +71,20 @@ public class IndexWork extends AbstractSimpleBulkableElasticsearchWork<Void> {
 
 		@Override
 		protected JsonObject buildBulkableActionMetadata() {
-			JsonObject index = new JsonObject();
-			index.addProperty( "_index", indexName.original );
-			index.addProperty( "_type", typeName.original );
-			index.addProperty( "_id", id.original );
+			JsonObject delete = new JsonObject();
+			delete.addProperty( "_index", indexName.original );
+			delete.addProperty( "_type", typeName.original );
+			delete.addProperty( "_id", id.original );
 
 			JsonObject result = new JsonObject();
-			result.add( "index", index );
+			result.add( "delete", delete );
 
 			return result;
 		}
 
 		@Override
-		public IndexWork build() {
-			return new IndexWork( this );
+		public DeleteWork build() {
+			return new DeleteWork( this );
 		}
 	}
 }
