+++ backend/elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/work/impl/IndexWork.java
@@ -4,27 +4,21 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.backend.elasticsearch.work.real.impl;
+package org.hibernate.search.backend.elasticsearch.work.impl;
 
 import org.hibernate.search.backend.elasticsearch.client.spi.ElasticsearchRequest;
 import org.hibernate.search.backend.elasticsearch.client.spi.ElasticsearchResponse;
 import org.hibernate.search.backend.elasticsearch.util.spi.URLEncodedString;
-import org.hibernate.search.backend.elasticsearch.work.builder.impl.DeleteWorkBuilder;
-import org.hibernate.search.backend.elasticsearch.work.impl.ElasticsearchWorkExecutionContext;
-import org.hibernate.search.backend.elasticsearch.work.real.accessor.impl.DefaultElasticsearchRequestSuccessAssessor;
-import org.hibernate.search.backend.elasticsearch.work.real.accessor.impl.ElasticsearchRequestSuccessAssessor;
+import org.hibernate.search.backend.elasticsearch.work.builder.impl.IndexWorkBuilder;
 
 import com.google.gson.JsonObject;
 
 /**
  * @author Yoann Rodiere
  */
-public class DeleteWork extends AbstractSimpleBulkableElasticsearchWork<Void> {
+public class IndexWork extends AbstractSimpleBulkableElasticsearchWork<Void> {
 
-	private static final ElasticsearchRequestSuccessAssessor SUCCESS_ASSESSOR =
-			DefaultElasticsearchRequestSuccessAssessor.builder().ignoreErrorStatuses( 404 ).build();
-
-	public DeleteWork(Builder builder) {
+	public IndexWork(Builder builder) {
 		super( builder );
 	}
 
@@ -40,30 +34,34 @@ public class DeleteWork extends AbstractSimpleBulkableElasticsearchWork<Void> {
 
 	public static class Builder
 			extends AbstractSimpleBulkableElasticsearchWork.Builder<Builder>
-			implements DeleteWorkBuilder {
+			implements IndexWorkBuilder {
 		private final URLEncodedString indexName;
 		private final URLEncodedString typeName;
 		private final URLEncodedString id;
 		private final String routingKey;
+		private final JsonObject document;
 
-		public Builder(URLEncodedString indexName, URLEncodedString typeName, URLEncodedString id, String routingKey) {
-			super( indexName, SUCCESS_ASSESSOR );
+		public Builder(URLEncodedString indexName, URLEncodedString typeName, URLEncodedString id, String routingKey, JsonObject document) {
+			super( indexName, DefaultElasticsearchRequestSuccessAssessor.INSTANCE );
 			this.indexName = indexName;
 			this.typeName = typeName;
 			this.id = id;
 			this.routingKey = routingKey;
+			this.document = document;
 		}
 
 		@Override
 		protected ElasticsearchRequest buildRequest() {
 			ElasticsearchRequest.Builder builder =
-					ElasticsearchRequest.delete()
+					ElasticsearchRequest.put()
 					.pathComponent( indexName )
 					.pathComponent( typeName )
-					.pathComponent( id );
+					.pathComponent( id )
 
-			// TODO avoid this param using a smart orchestrator
-			builder.param( "refresh", true );
+					// TODO avoid this param using a smart orchestrator
+					.param( "refresh", true )
+
+					.body( document );
 
 			if ( routingKey != null ) {
 				builder.param( "routing", routingKey );
@@ -74,20 +72,20 @@ public class DeleteWork extends AbstractSimpleBulkableElasticsearchWork<Void> {
 
 		@Override
 		protected JsonObject buildBulkableActionMetadata() {
-			JsonObject delete = new JsonObject();
-			delete.addProperty( "_index", indexName.original );
-			delete.addProperty( "_type", typeName.original );
-			delete.addProperty( "_id", id.original );
+			JsonObject index = new JsonObject();
+			index.addProperty( "_index", indexName.original );
+			index.addProperty( "_type", typeName.original );
+			index.addProperty( "_id", id.original );
 
 			JsonObject result = new JsonObject();
-			result.add( "delete", delete );
+			result.add( "index", index );
 
 			return result;
 		}
 
 		@Override
-		public DeleteWork build() {
-			return new DeleteWork( this );
+		public IndexWork build() {
+			return new IndexWork( this );
 		}
 	}
 }
