+++ backend/elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/work/impl/OpenIndexWork.java
@@ -4,19 +4,20 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.backend.elasticsearch.work.real.impl;
+package org.hibernate.search.backend.elasticsearch.work.impl;
 
 import org.hibernate.search.backend.elasticsearch.client.spi.ElasticsearchRequest;
 import org.hibernate.search.backend.elasticsearch.client.spi.ElasticsearchResponse;
 import org.hibernate.search.backend.elasticsearch.client.impl.Paths;
 import org.hibernate.search.backend.elasticsearch.util.spi.URLEncodedString;
-import org.hibernate.search.backend.elasticsearch.work.builder.impl.CloseIndexWorkBuilder;
-import org.hibernate.search.backend.elasticsearch.work.impl.ElasticsearchWorkExecutionContext;
-import org.hibernate.search.backend.elasticsearch.work.real.accessor.impl.DefaultElasticsearchRequestSuccessAssessor;
+import org.hibernate.search.backend.elasticsearch.work.builder.impl.OpenIndexWorkBuilder;
 
-public class CloseIndexWork extends AbstractSimpleElasticsearchWork<Void> {
+/**
+ * @author Yoann Rodiere
+ */
+public class OpenIndexWork extends AbstractSimpleElasticsearchWork<Void> {
 
-	protected CloseIndexWork(Builder builder) {
+	protected OpenIndexWork(Builder builder) {
 		super( builder );
 	}
 
@@ -27,8 +28,7 @@ public class CloseIndexWork extends AbstractSimpleElasticsearchWork<Void> {
 
 	public static class Builder
 			extends AbstractSimpleElasticsearchWork.Builder<Builder>
-			implements CloseIndexWorkBuilder {
-
+			implements OpenIndexWorkBuilder {
 		private final URLEncodedString indexName;
 
 		public Builder(URLEncodedString indexName) {
@@ -41,14 +41,14 @@ public class CloseIndexWork extends AbstractSimpleElasticsearchWork<Void> {
 			ElasticsearchRequest.Builder builder =
 					ElasticsearchRequest.post()
 					.pathComponent( indexName )
-					.pathComponent( Paths._CLOSE );
+					.pathComponent( Paths._OPEN );
 
 			return builder.build();
 		}
 
 		@Override
-		public CloseIndexWork build() {
-			return new CloseIndexWork( this );
+		public OpenIndexWork build() {
+			return new OpenIndexWork( this );
 		}
 	}
 }
