+++ backend/elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/work/impl/OptimizeWork.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.backend.elasticsearch.work.real.impl;
+package org.hibernate.search.backend.elasticsearch.work.impl;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -13,16 +13,18 @@ import org.hibernate.search.backend.elasticsearch.client.spi.ElasticsearchReques
 import org.hibernate.search.backend.elasticsearch.client.spi.ElasticsearchResponse;
 import org.hibernate.search.backend.elasticsearch.client.impl.Paths;
 import org.hibernate.search.backend.elasticsearch.util.spi.URLEncodedString;
-import org.hibernate.search.backend.elasticsearch.work.builder.impl.RefreshWorkBuilder;
-import org.hibernate.search.backend.elasticsearch.work.impl.ElasticsearchWorkExecutionContext;
-import org.hibernate.search.backend.elasticsearch.work.real.accessor.impl.DefaultElasticsearchRequestSuccessAssessor;
+import org.hibernate.search.backend.elasticsearch.work.builder.impl.OptimizeWorkBuilder;
 
 /**
+ * An optimize work for ES5, using the ForceMerge API.
+ * <p>
+ * The ForceMerge API replaces the removed Optimize API in ES5.
+ *
  * @author Yoann Rodiere
  */
-public class RefreshWork extends AbstractSimpleElasticsearchWork<Void> {
+public class OptimizeWork extends AbstractSimpleElasticsearchWork<Void> {
 
-	protected RefreshWork(Builder builder) {
+	protected OptimizeWork(Builder builder) {
 		super( builder );
 	}
 
@@ -33,7 +35,7 @@ public class RefreshWork extends AbstractSimpleElasticsearchWork<Void> {
 
 	public static class Builder
 			extends AbstractSimpleElasticsearchWork.Builder<Builder>
-			implements RefreshWorkBuilder {
+			implements OptimizeWorkBuilder {
 		private List<URLEncodedString> indexNames = new ArrayList<>();
 
 		public Builder() {
@@ -42,7 +44,7 @@ public class RefreshWork extends AbstractSimpleElasticsearchWork<Void> {
 
 		@Override
 		public Builder index(URLEncodedString indexName) {
-			indexNames.add( indexName );
+			this.indexNames.add( indexName );
 			return this;
 		}
 
@@ -55,14 +57,14 @@ public class RefreshWork extends AbstractSimpleElasticsearchWork<Void> {
 				builder.multiValuedPathComponent( indexNames );
 			}
 
-			builder.pathComponent( Paths._REFRESH );
+			builder.pathComponent( Paths._FORCEMERGE );
 
 			return builder.build();
 		}
 
 		@Override
-		public RefreshWork build() {
-			return new RefreshWork( this );
+		public OptimizeWork build() {
+			return new OptimizeWork( this );
 		}
 	}
 }
