+++ backend/elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/work/impl/RefreshWork.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.backend.elasticsearch.work.real.impl;
+package org.hibernate.search.backend.elasticsearch.work.impl;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -13,20 +13,14 @@ import org.hibernate.search.backend.elasticsearch.client.spi.ElasticsearchReques
 import org.hibernate.search.backend.elasticsearch.client.spi.ElasticsearchResponse;
 import org.hibernate.search.backend.elasticsearch.client.impl.Paths;
 import org.hibernate.search.backend.elasticsearch.util.spi.URLEncodedString;
-import org.hibernate.search.backend.elasticsearch.work.builder.impl.OptimizeWorkBuilder;
-import org.hibernate.search.backend.elasticsearch.work.impl.ElasticsearchWorkExecutionContext;
-import org.hibernate.search.backend.elasticsearch.work.real.accessor.impl.DefaultElasticsearchRequestSuccessAssessor;
+import org.hibernate.search.backend.elasticsearch.work.builder.impl.RefreshWorkBuilder;
 
 /**
- * An optimize work for ES5, using the ForceMerge API.
- * <p>
- * The ForceMerge API replaces the removed Optimize API in ES5.
- *
  * @author Yoann Rodiere
  */
-public class OptimizeWork extends AbstractSimpleElasticsearchWork<Void> {
+public class RefreshWork extends AbstractSimpleElasticsearchWork<Void> {
 
-	protected OptimizeWork(Builder builder) {
+	protected RefreshWork(Builder builder) {
 		super( builder );
 	}
 
@@ -37,7 +31,7 @@ public class OptimizeWork extends AbstractSimpleElasticsearchWork<Void> {
 
 	public static class Builder
 			extends AbstractSimpleElasticsearchWork.Builder<Builder>
-			implements OptimizeWorkBuilder {
+			implements RefreshWorkBuilder {
 		private List<URLEncodedString> indexNames = new ArrayList<>();
 
 		public Builder() {
@@ -46,7 +40,7 @@ public class OptimizeWork extends AbstractSimpleElasticsearchWork<Void> {
 
 		@Override
 		public Builder index(URLEncodedString indexName) {
-			this.indexNames.add( indexName );
+			indexNames.add( indexName );
 			return this;
 		}
 
@@ -59,14 +53,14 @@ public class OptimizeWork extends AbstractSimpleElasticsearchWork<Void> {
 				builder.multiValuedPathComponent( indexNames );
 			}
 
-			builder.pathComponent( Paths._FORCEMERGE );
+			builder.pathComponent( Paths._REFRESH );
 
 			return builder.build();
 		}
 
 		@Override
-		public OptimizeWork build() {
-			return new OptimizeWork( this );
+		public RefreshWork build() {
+			return new RefreshWork( this );
 		}
 	}
 }
